+++ ffi/miri-tests/mock-redis/src/lib.rs
@@ -133,3 +133,13 @@ impl PipelineRetryStrategy {
         PipelineRetryStrategy
     }
 }
+
+pub fn parse_redis_url(input: &str) -> Option<url::Url> {
+    match url::Url::parse(input) {
+        Ok(result) => match result.scheme() {
+            "redis" | "rediss" | "redis+unix" | "unix" => Some(result),
+            _ => None,
+        },
+        Err(_) => None,
+    }
+}